% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.hmc}
\alias{coef.hmc}
\title{Extract Coefficients from an HMC Model
This function extracts coefficients from an HMC model object.}
\usage{
\method{coef}{hmc}(object, predictor = c("location", "scale"), ...)
}
\arguments{
\item{object}{An HMC model object.}

\item{predictor}{A character vector specifying the predictor(s) for which coefficients are extracted.
Options include "location" and "scale." Default is c("location", "scale"), which returns both location and scale coefficients.}

\item{...}{Additional arguments (not used).}
}
\value{
A list of coefficients corresponding to the specified predictor(s).
}
\description{
Extract Coefficients from an HMC Model
This function extracts coefficients from an HMC model object.
}
