% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc}
\alias{hmc}
\title{Hamiltonian Markov Chain Monte Carlo (HMC) Sampling}
\usage{
hmc(
  location = y ~ x,
  scale = ~1,
  data = environment(location),
  light = FALSE,
  num_samples = 1000,
  verbose = TRUE,
  threshold = 1/10000,
  include_warmup = FALSE,
  dual_average = list(DELTA = 0.65, LAMBDA = 5, KAPPA = 0.75, GAMMA = 0.05, t0 = 10),
  num_adapt = 500,
  num_warmup = 500,
  max_L = 250
)
}
\arguments{
\item{location}{A formula specifying the model's location component (e.g., y ~ x).}

\item{scale}{A formula specifying the model's scale component (e.g., ~ 1).}

\item{data}{The dataset containing the model's variables.}

\item{light}{A logical indicating whether to use the "light" version of the model (default is FALSE).}

\item{num_samples}{The number of MCMC samples to generate (default is 1000).}

\item{verbose}{A logical indicating whether to display progress messages (default is TRUE).}

\item{threshold}{The convergence threshold for adaptive step size tuning (default is 1/10000).}

\item{include_warmup}{A logical indicating whether to include warm-up samples in the final results (default is FALSE).}

\item{dual_average}{A list of parameters for dual averaging step size adaptation.}

\item{num_adapt}{The maximum number of adaptation iterations (default is 500).}

\item{num_warmup}{The number of warm-up iterations (default is 500).}

\item{max_L}{The maximum number of leapfrog steps for HMC (default is 250).}
}
\value{
An HMC model object containing MCMC samples and results.
}
\description{
This function performs HMC sampling for a given location-scale regression model.
}
\examples{
library(gamlss.data)
data(abdom, package = "gamlss.data")
output <- hmc(y ~ x, ~x, data = abdom)

}
