% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.hmc}
\alias{logLik.hmc}
\title{Extract Log-Likelihood from an HMC Model
This function extracts the log-likelihood from an HMC model object.}
\usage{
\method{logLik}{hmc}(object, ...)
}
\arguments{
\item{object}{An HMC model object.}

\item{...}{Additional arguments (not used).}
}
\description{
Extract Log-Likelihood from an HMC Model
This function extracts the log-likelihood from an HMC model object.
}
