% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.hmc}
\alias{plot.hmc}
\title{Plot Method for an HMC Model}
\usage{
\method{plot}{hmc}(
  x,
  predictor = c("location", "scale"),
  type = c("trace", "hist", "density", "acf", "roll_avg"),
  exclude_warmup = TRUE,
  ...
)
}
\arguments{
\item{x}{An HMC model object.}

\item{predictor}{A character vector specifying the predictor for which plots are generated.
Options include "location" and "scale." Default is c("location", "scale"), which generates plots for both location and scale predictors.}

\item{type}{A character vector specifying the type of plots to generate. Options include "trace," "hist," "density," "acf," and "roll_avg."}

\item{exclude_warmup}{A logical value indicating whether to exclude warm-up iterations when generating plots. Default is TRUE.}

\item{...}{Additional arguments (not used).}
}
\description{
This function generates various types of plots for an HMC model object.
}
