% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.hmc}
\alias{print.hmc}
\title{Print Method for an HMC Model}
\usage{
\method{print}{hmc}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An HMC model object.}

\item{digits}{The number of digits to be used for printing numeric values. Default is max(3, getOption("digits") - 3).}

\item{...}{Additional arguments (not used).}
}
\description{
This function provides a customized print method for an HMC model object.
}
