% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.hmc}
\alias{summary.hmc}
\title{Summary Method for an HMC Model}
\usage{
\method{summary}{hmc}(object, type = "hmc", digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{An HMC model object.}

\item{type}{A character string specifying the type of summary to generate. Default is 'hmc'.}

\item{digits}{The number of digits to be used for printing numeric values. Default is max(3, getOption("digits") - 3).}

\item{...}{Additional arguments (not used).}
}
\value{
A summary of the HMC model object.
}
\description{
This function provides a summary of an HMC model object.
}
